[yeelight] Made constants and handler packages internal (#4036)
