+++ addons/binding/org.openhab.binding.yeelight/META-INF/MANIFEST.MF
@@ -17,9 +17,5 @@ Import-Package:
  org.eclipse.smarthome.core.thing.binding.builder,
  org.eclipse.smarthome.core.thing.type,
  org.eclipse.smarthome.core.types,
- org.openhab.binding.yeelight,
- org.openhab.binding.yeelight.handler,
  org.slf4j
 Service-Component: OSGI-INF/*.xml
-Export-Package: org.openhab.binding.yeelight,
- org.openhab.binding.yeelight.handler
